Cambia ids externas por bigints

+++ db/migrate/20170820181939_create_openings.rb
@@ -2,9 +2,9 @@ class CreateOpenings < ActiveRecord::Migration[5.1]
   def change
     create_table :openings do |t|
       t.string :name, null: false
-      t.integer :service_id, null: false
-      t.integer :test_stage_id, null: false
-      t.integer :passed_stage_id, null: false
+      t.bigint :service_id, null: false
+      t.bigint :test_stage_id, null: false
+      t.bigint :passed_stage_id, null: false
       t.timestamps
 
       t.index :service_id, unique: true
