+++ db/schema.rb
@@ -19,7 +19,7 @@ ActiveRecord::Schema.define(version: 20170820230159) do
     t.string "email", null: false
     t.string "first_name"
     t.string "last_name"
-    t.integer "service_id", null: false
+    t.bigint "service_id", null: false
     t.bigint "opening_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
@@ -30,9 +30,9 @@ ActiveRecord::Schema.define(version: 20170820230159) do
 
   create_table "openings", force: :cascade do |t|
     t.string "name", null: false
-    t.integer "service_id", null: false
-    t.integer "test_stage_id", null: false
-    t.integer "passed_stage_id", null: false
+    t.bigint "service_id", null: false
+    t.bigint "test_stage_id", null: false
+    t.bigint "passed_stage_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["service_id"], name: "index_openings_on_service_id", unique: true
